requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# AI/LLM
openai>=1.0.0

# HTTP requests
requests>=2.31.0

# Text extraction (Windows compatible versions)
docx2txt>=0.8
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=0.8.11

# Alternative PDF processing
pdfplumber>=0.9.0

# Utilities
pathlib2>=2.3.7